Prevent placement of fire into nodes above water
This addresses the edge case where fire is set to the top of a non-water
block that represents water (such as kelp).

Note this is forbidding theoretically legitimate setups for fire over
water where fire is set to the side of a block diagonally from the
water.

Fire is still permitted next to water (so it can be set to the face of a
block diagonally from a water column).