Remove ugly workaround for death formspec
kay27 added this because if you sleep and blow up by TNT the player gets
stuck in a ghost state were they can move around but not interact with
anything until they quit and log in again. This is due to an engine bug
were if a player dies when they are in a formspec, the death message
does not show up.

It does however not matter how the player died, this bug is always
triggered. Because this bug can just as easily be triggered by being
killed by a mob or another player it does not make sense to have such an
ugly workaround for a specific mod.