Shulker: lower shoot interval to 1 second
In this video we can see the shulkers are shooting at a rate of once per second.
https://youtube.com/shorts/8tUFaSZ4b7I?si=sszp4cnyXR4dXg9B

The wiki confirms this but mentions a range of 1 to 5.5 seconds.
I think this range is caused by the shulker closing and stopping shots for a few seconds but no longer than the max interval of 5.5

https://minecraft.fandom.com/wiki/Shulker#:~:text=The%20shulker%20continues%20firing%20every,unaffected%20by%20Levitation%20when%20hit