ENTITIES/mcl_mobs: fix broken passive mob timer
The timer has to be decremented by the full time elapsed since the last
call, but check_timer isn't called on every invocation of the globalstep
function and therefore misses time. Decrement it directly in the
globalstep function instead.

Previously check_timer seems to have assumed to be called with the
spawn_class as parameter (called mob_type, though), but actually it is
called on the spawn definition, leading to always returning true.
Instead retrieve spawn class from mob def found in registry.

The timer reset logic was reset even on non passive mobs, erronously
reducing the chance of spawning passive mobs. (Not noticable because
logic couldn't identify passive mobs correctly.)

Also PASSIVE_INTERVAL had been set to 200 instead of the 20 seconds
documented in the minecraft wiki. Change that, too.