Make villager trade interface scrollbar usable on mobile devices
Since minetest doesn't have a scroll gesture for mobile devices (yet), the
scrollbars have to be usable. This requires that the scrollbar is not
overlapped by any container and has a reasonable size.

Therefore shrink the scroll container in the villager trading interface to
no longer overlap the scrollbar and increase the scrollbar width.