Relax mcl_util.is_pointing_above_middle for touch users
Using the touch interface without crosshair the player may not actually look
directly at the node/object, so player:get_look_dir() cannot be used to infer
the location of the touch interaction. But it is good enough to check the
general look direction to allow touch users to somewhat reproducibly affect
node placement of e.g. slabs and stairs.

Also adds a probably too extensive comment explaining the reasoning:-)

This change should be a noop for crosshair users.